moved the bulksending to mailtransport
